@@ -5,46 +5,50 @@ publish_to: 'none'
 version: 2.3+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  sqflite: ^1.1.7+2
-  sqfentity: ^1.2.3
-  sqfentity_gen: ^1.2.3
+  sqflite: ^2.0.0+3
+  sqfentity: ^2.0.0
+  sqfentity_gen: ^2.0.0
   path: ^1.8.0
 
-  shared_preferences: ^2.0.5
+  shared_preferences: ^2.0.6
   intl: ^0.17.0
-  connectivity: ^3.0.3
+  connectivity: ^3.0.6
   diacritic: ^0.1.3
-  wakelock: ^0.5.1
+  wakelock: ^0.5.2
     
-  provider: ^5.0.0
+  provider: ^6.0.0
 
   modal_bottom_sheet: ^2.0.0
 
-  url_launcher: ^6.0.3
-  flutter_share: ^2.0.0
+  url_launcher: ^6.0.6
+  share_plus: ^2.1.4
 
   youtube_player_flutter: ^8.0.0
-  flutter_svg: ^0.21.0+1
+  flutter_svg: ^0.22.0
   flutter_reorderable_list: ^1.0.0
   flutter_staggered_grid_view: ^0.4.0
 
-  firebase_analytics: ^7.1.1
+  firebase_analytics: ^8.1.1
 
   uni_links: ^0.5.1
 
   flutter_nearby_connections: ^1.0.11
-  device_info: ^2.0.0
+  device_info: ^2.0.2
 
-  cupertino_icons: ^1.0.2
+  auto_size_text: ^3.0.0-nullsafety.0
+
+  fluttertoast: ^8.0.8
+
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
-  build_runner: any
+  build_runner: ^2.0.4
 
 flutter:
   uses-material-design: true
@@ -57,6 +61,10 @@ flutter:
     - family: Amiri
       fonts:
         - asset: assets/fonts/Amiri-Regular.ttf
+
+    - family: KaiseiHarunoUmi
+      fonts:
+        - asset: assets/fonts/KaiseiHarunoUmi-Regular.ttf
 
     - family: Nunito
       fonts:
