@@ -1,87 +1,87 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:zpevnik/theme.dart';
 import 'package:zpevnik/platform/mixin.dart';
+import 'package:zpevnik/screens/utils/updateable.dart';
+import 'package:zpevnik/theme.dart';
 
-class PlatformDialog extends StatelessWidget with PlatformWidgetMixin {
+Future<T?> showPlatformDialog<T>(BuildContext context, Widget Function(BuildContext) builder) {
+  if (Theme.of(context).platform == TargetPlatform.iOS)
+    return showCupertinoDialog<T>(context: context, barrierDismissible: false, builder: builder);
+  else
+    return showDialog<T>(context: context, barrierDismissible: false, builder: builder);
+}
+
+class PlatformDialog extends StatefulWidget {
   final String title;
   final String initialValue;
   final String submitText;
-  final Function(String) onSubmit;
 
-  const PlatformDialog({Key key, this.title, this.initialValue, this.onSubmit, this.submitText = 'Vytvořit'})
-      : super(key: key);
+  const PlatformDialog({
+    Key? key,
+    this.title = '',
+    this.initialValue = '',
+    this.submitText = '',
+  }) : super(key: key);
 
   @override
-  Widget androidWidget(BuildContext context) {
-    final textFieldController = TextEditingController()..text = initialValue;
+  _PlatformDialogState createState() => _PlatformDialogState();
+}
 
+class _PlatformDialogState extends State<PlatformDialog> with PlatformMixin, Updateable {
+  final _textFieldController = TextEditingController();
+
+  @override
+  void initState() {
+    super.initState();
+
+    _textFieldController.text = widget.initialValue;
+  }
+
+  @override
+  Widget buildAndroid(BuildContext context) {
     return AlertDialog(
-      title: Text(title),
-      content: Container(
-        child: TextField(
-          decoration: InputDecoration(border: InputBorder.none, hintText: 'Název'),
-          controller: textFieldController,
-        ),
+      title: Text(widget.title),
+      content: TextField(
+        autofocus: true,
+        decoration: InputDecoration(border: InputBorder.none, hintText: 'Název'),
+        controller: _textFieldController,
       ),
-      actions: [
-        TextButton(
-          child: Text('Zrušit', style: AppTheme.of(context).bodyTextStyle.copyWith(color: Colors.red)),
-          onPressed: () => Navigator.of(context).pop(),
-        ),
-        // fixme: don't know better way to do it now, but there must be
-        ChangeNotifierProvider.value(
-          value: textFieldController,
-          child: Consumer<TextEditingController>(
-            builder: (context, controller, _) => TextButton(
-              child: Text(submitText),
-              onPressed: controller.text.isEmpty
-                  ? null
-                  : () {
-                      onSubmit(controller.text);
-                      Navigator.of(context).pop();
-                    },
-            ),
-          ),
-        ),
-      ],
+      actions: _actions(context),
     );
   }
 
   @override
-  Widget iOSWidget(BuildContext context) {
-    final textFieldController = TextEditingController()..text = initialValue;
-
-    return CupertinoTheme(
-      data: AppTheme.of(context).cupertinoTheme, // fixme: is set again, because it is lost somewhere
-      child: CupertinoAlertDialog(
-        title: Text(title),
-        content: Container(child: CupertinoTextField(placeholder: 'Název', controller: textFieldController)),
-        actions: [
-          TextButton(
-            child: Text('Zrušit', style: AppTheme.of(context).bodyTextStyle.copyWith(color: Colors.red)),
-            onPressed: () => Navigator.of(context).pop(),
-          ),
-          // fixme: don't know better way to do it now, but there must be
-          ChangeNotifierProvider.value(
-            value: textFieldController,
-            child: Consumer<TextEditingController>(
-              builder: (context, controller, _) => TextButton(
-                child: Text(submitText,
-                    style: AppTheme.of(context).bodyTextStyle.copyWith(
-                        color: controller.text.isEmpty ? AppTheme.of(context).textColor.withAlpha(0x77) : null)),
-                onPressed: controller.text.isEmpty
-                    ? null
-                    : () {
-                        onSubmit(controller.text);
-                        Navigator.of(context).pop();
-                      },
-              ),
-            ),
-          ),
-        ],
+  Widget buildIos(BuildContext context) {
+    return CupertinoAlertDialog(
+      title: Text(widget.title),
+      content: CupertinoTextField(
+        autofocus: true,
+        placeholder: 'Název',
+        controller: _textFieldController,
       ),
+      actions: _actions(context),
     );
   }
+
+  List<Widget> _actions(BuildContext context) {
+    final appTheme = AppTheme.of(context);
+
+    final confirmTextStyle = appTheme.bodyTextStyle
+        ?.copyWith(color: _textFieldController.text.isEmpty ? appTheme.textColor.withAlpha(0x77) : null);
+
+    return [
+      TextButton(
+        child: Text('Zrušit', style: appTheme.bodyTextStyle?.copyWith(color: Colors.red)),
+        onPressed: () => Navigator.of(context).pop(),
+      ),
+      TextButton(
+        child: Text(widget.submitText, style: confirmTextStyle),
+        onPressed:
+            _textFieldController.text.isEmpty ? null : () => Navigator.of(context).pop(_textFieldController.text),
+      ),
+    ];
+  }
+
+  @override
+  List<Listenable> get listenables => [_textFieldController];
 }
