@@ -2,15 +2,14 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:zpevnik/platform/mixin.dart';
 
-class PlatformSwitch extends StatelessWidget with PlatformWidgetMixin {
+class PlatformSwitch extends StatelessWidget with PlatformMixin {
   final bool value;
-  final Function(bool) onChanged;
+  final Function(bool)? onChanged;
 
-  const PlatformSwitch({Key key, this.value, this.onChanged}) : super(key: key);
-
-  @override
-  Widget androidWidget(BuildContext context) => Switch(value: value, onChanged: onChanged);
+  const PlatformSwitch({Key? key, required this.value, this.onChanged}) : super(key: key);
 
   @override
-  Widget iOSWidget(BuildContext context) => CupertinoSwitch(value: value, onChanged: onChanged);
+  Widget buildAndroid(BuildContext context) => Switch(value: value, onChanged: onChanged);
+  @override
+  Widget buildIos(BuildContext context) => CupertinoSwitch(value: value, onChanged: onChanged);
 }
