@@ -1,14 +1,13 @@
 import 'package:flutter/material.dart';
 
-// from https://stackoverflow.com/a/53905427/8592888
-class HexColor extends Color {
-  static int _getColorFromHex(String hexColor) {
-    hexColor = hexColor.toUpperCase().replaceAll("#", "");
-    if (hexColor.length == 6) {
-      hexColor = "FF" + hexColor;
-    }
-    return int.parse(hexColor, radix: 16);
-  }
+extension HexColor on Color {
+  static Color? fromHex(String? hexColor) {
+    if (hexColor == null) return null;
 
-  HexColor(final String hexColor) : super(_getColorFromHex(hexColor));
+    hexColor = hexColor.toUpperCase().replaceAll("#", "");
+
+    if (hexColor.length == 6) hexColor = "FF" + hexColor;
+
+    return Color(int.parse(hexColor, radix: 16));
+  }
 }
